put solara app in the layout (will work with current cosmicds baselayout too)
